add make theme component
