+++ packages/ui/src/index.js
@@ -27,3 +27,4 @@ export { default as Textarea } from '../packages/Textarea/src/index';
 export { default as TextWrap } from '../packages/TextWrap/src/index';
 export { default as Toggle } from '../packages/Toggle/src/index';
 export { default as Tooltip } from '../packages/Tooltip/src/index';
+export * from '../packages/Theme/src/index';
